+++ lib/src/web/web_instrumentation_stub.dart
@@ -6,6 +6,8 @@ import 'web_instrumentation_options.dart';
 
 bool isBotTraffic() => false;
 
+String? pagePath() => null;
+
 Map<String, Object> browserResourceAttributes() => const {};
 
 void enable(